Move KeyProvider class into sync-core
The KeyProvider class is a class which could be used on either
Android or JavaSE. While we only currently provide SQLCipher on
Android, we plan to offer it on all platforms that both we and
SQLCipher support.
