+++ sync-android/src/main/java/com/cloudant/sync/sqlite/android/AndroidSQLCipherSQLite.java
@@ -28,7 +28,7 @@ package com.cloudant.sync.sqlite.android;
 import com.cloudant.sync.sqlite.ContentValues;
 import com.cloudant.sync.sqlite.Cursor;
 import com.cloudant.sync.sqlite.SQLDatabase;
-import com.cloudant.sync.sqlite.android.encryption.KeyProvider;
+import com.cloudant.sync.datastore.encryption.KeyProvider;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
